backend/app/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
python-dotenv
openpyxl